feat: Adicionando EnderecoRepository, implementando EnderecoServico, alterando IRepository e Endereco

+++ backend/src/com/dbc/repository/IRepository.java
@@ -1,7 +1,6 @@
 package com.dbc.repository;
 
 import com.dbc.exceptions.BancoDeDadosException;
-import model.entidades.Agenda;
 
 import java.sql.Connection;
 import java.sql.SQLException;
@@ -9,7 +8,7 @@ import java.util.List;
 
 public interface IRepository<CHAVE, OBJETO> {
 
-    Integer getProximoId(Connection connection) throws SQLException;
+    Long getProximoId(Connection connection) throws SQLException;
     OBJETO cadastrar(OBJETO objeto) throws BancoDeDadosException;
     List<OBJETO> listar() throws BancoDeDadosException;
     OBJETO listarUm(CHAVE id) throws BancoDeDadosException;
